requirements.txt
google-adk>=1.12.0
litellm
python-dotenv
IPython